+++ src/core/api/index.ts
@@ -1,3 +1,2 @@
 /* istanbul ignore file */
-export * from './auth';
 export * from './types';