+++ src/core/auth/index.ts
@@ -1,0 +1,4 @@
+/* istanbul ignore file */
+export * from './context';
+export * from './utils';
+export { default as Auth, CognitoHostedUIIdentityProvider } from '@aws-amplify/auth';