+++ src/screens/Dashboard/Events/scenes/EmailConfirm/index.tsx
@@ -1,5 +1,5 @@
 import { useSnackbar } from 'components/Feedback';
-import { useAuth } from 'core/api';
+import { useAuth } from 'core/auth';
 import React, { FC, useState } from 'react';
 import { StyleProps } from 'styles';
 import { Log } from 'utils';