+++ src/screens/Dashboard/index.tsx
@@ -1,7 +1,7 @@
 import { Grid, makeStyles, Theme, useTheme } from '@material-ui/core';
 import { ScreenTitle, View } from 'components/Common';
 import { DashboardAppBar, DashboardMobileMenu, DashboardAppBarBtn } from 'components/Dashboard';
-import { useAuth } from 'core/api';
+import { useAuth } from 'core/auth';
 import React, { FC, useEffect, useState } from 'react';
 import { Redirect, Route, Switch, useHistory } from 'react-router-dom';
 import { routes } from 'screens/consts';